requirements.txt
paho-mqtt
jsonpath-rw
jsonpath-rw-ext